Extracting Impl definitions to header
